pre-save middleware ownerID !== partnerID. separated online and offline
